Refactor baggage controller and route
This commit refactors the `getBaggagesByUser` function in the `baggage.controller.js` file to use `req.params` instead of `req.body` for retrieving the user parameter. It also adds a new `deleteBaggage` function to handle deleting baggage based on `tagData`. The `baggage.route.js` file is updated to include the new route for deleting baggage.